sh: Avoid setenv after fork
setenv() is not async-signal-safe and as such should not be used
between fork/exec of a multi-threaded app: if one thread is
manipulating the current environment (which may entail obtaining a
malloc() mutex) when another thread calls fork(), the resulting
child's attempt to use setenv() could deadlock or see a broken environ
because the thread owning the lock no longer exists to release it.
Besides, it is more efficient to update the environment exactly once
in the parent, rather than after every fork(), since .load gets run
exactly once.  If we are worried about the environment change in sh
affecting what filters may do, then we'd have to compute an
alternative environ in the parent and switch the child process to
using execvpe(); but since filters are in-tree, it's easy to audit
that none of them care about the change.

While at it, check for (unlikely) failure of setenv.

Signed-off-by: Eric Blake <eblake@redhat.com>
(cherry picked from commit 4c588cb415610bf8526f0ec881a8fd6a0d769f04)