plugins: Match docs for .errno_is_preserved
Ever since commit 69ae137f, the docs claimed that we check
.errno_is_preserved == 1, but the code has checked for != 0.  It's
easy enough to add new fields if we need new semantics, so there is no
reason to break any existing plugin that was setting the field to 2;
so change the docs to match reality, but also change --dump-plugin to
canonicalize its output.

Furthermore, the mention of .errno_is_preserved in the docs was rather
hidden; a new section will make it easier to add future knobs that
likewise affect the plugin as a whole and not an individual
connection.

Fixes: 69ae137f
Signed-off-by: Eric Blake <eblake@redhat.com>