plugins: Expose more thread_model details in --dump-plugin
When we added .thread_model to plugins, we forgot to expose
has_thread_model in the --dump-plugin output.  What's more, it is
useful to determine if the plugin itself is aware of tighter
restrictions imposed by dynamic choices, by outputting both
'max_thread_model' (what the plugin was compiled as) and
'thread_model' (the resulting model based on runtime knowledge).

Note that this means that a plugin's .thread_model can be called prior
to .config_complete, so the sh plugin must be prepared for a missing
script.  In fact, exposing additional information during --dump-plugin
will make it easier for a future patch to make sh actually allow
PARALLEL, and still have introspection on which model will actually be
used.

Fixes: 9911a9d3
Signed-off-by: Eric Blake <eblake@redhat.com>