sh: Enable parallel thread model, when possible
When we first created the sh plugin, we copied the thread model used
in most other language bindings of SERIALIZE_ALL_REQUESTS, because it
was easier to reason about and there was no way for a script to
override our choice.  However, we've since added support for scripts
to request a tighter model when needed, and we've also just fixed
remaining bugs about fds inadvertently leaked into the shell script
(at least, when pipe2 and accept4 are supported).  And the shell
itself is sufficiently expressive (even if not necessarily the most
efficient) for two scripts to come up with ways to enforce mutual
exclusion (for example, a script can use the success or failure of
'mkdir' on a coordinated filename to learn if it is first past the
gate, or set up FIFOs for interprocess communication).  Still, for
backwards-compatibility, this has to be something that the shell
script opts in to, rather than changing the default.

Many existing uses of shell scripts are tweaked to demonstrate the use
of opting in to parallel, and test-dump-plugin is massaged to show
both raising and lowering the thread model, modulo a capped maximum on
Haiku due to atomic CLOEXEC problems.  Meanwhile, the new
test-parallel-sh.sh test copies heavily from test-parallel-file.sh.

Signed-off-by: Eric Blake <eblake@redhat.com>