plugins: Permit ENOTSUP as synonym for EOPNOTSUPP
POSIX allows but does not require ENOTSUP and EOPNOTSUPP to map to the
same errno value.  This patch has no impact on Linux, but does affect
other systems: we want to be permissive in what we accept (either
spelling, since the two are commonly confused), while strict in what
we generate (we documented EOPNOTSUPP as the trigger to fall back to
.write, so stick to that internally where it makes sense).  An
upcoming NBD protocol extension is proposing the exposure of an
NBD_ENOTSUP error, and so we'll need to make sure that both errno
values (when they are distinct) feed into that single wire value.

Signed-off-by: Eric Blake <eblake@redhat.com>