server: Assert sane error responses
Our filters document that calling through next_ops should reliably set
*err on failure, and in turn that the filter must set *err if
returning -1.  Since plugins.c wrapper ensures that *err is always set
regardless of the plugin, the only place where *err could be left
unset is in a broken filter.  Our documentation also states that a
filter must never observe or return EOPNOTSUPP from a failed .zero,
since any fallback to .pwrite should have already happened (this
restriction will be changed in future patches when the new
NBD_CMD_FLAG_FAST_ZERO is added, but for now it is worth making sure
we obey).  Since filters are always in-tree, it is appropriate to
assert that none of our backends are forgetting to set *err correctly.

Signed-off-by: Eric Blake <eblake@redhat.com>