nozero: More efficient FUA handling
The nozero filter can split a large request with FUA into several
smaller requests; optimize whether the FUA flag is passed on to the
next layer based on whether FUA is emulated with flush (only the last
write needs it) or is natively supported (every write needs it).
Missed in commit df0cc21d.

Signed-off-by: Eric Blake <eblake@redhat.com>