sh: Fix .zero falling back to .pwrite
We weren't actually testing the sh plugin's ability to have .zero fall
back to .pwrite.  Adding a test for this flushed out a bug in commit
95a72f5d: setting *err only if it is not already set only picks up a
.pwrite error if we clear *err after a failed .zero.  Also, it's worth
documenting the tricks used in this test: a .zero implementation can
prefix something to stderr whether or not it eventually fails, which
is a lot nicer than redirecting stderr into temporary storage then
replaying it later (necessary since the sh detection of ENOTSUP
requires a specific prefix).

Signed-off-by: Eric Blake <eblake@redhat.com>