tests: Increase timeouts on test-shutdown.sh
Testing that shutdown happens in a timely manner is inherently a racy
aspect (it's hard to guarantee that a slow heavily-loaded machine will
react as fast as desired.  Rich hit a test failure when running 'make
-j16', complaining "Unexpected status; nbdkit didn't react fast enough
to signal".

Improve the test by lengthening the delay of the server (so we have
more time to react to other timing events even under load),
lengthening the timeout of the 'quit' command that normally completes
quickly without load, and rewriting the 'kill -0' test for a completed
server into a longer timeout via a loop (to let the test remain
responsive and complete quickly on a machine without load, but to
allow longer before failing the test on a machine with load).  On my
machine, the test still completes in under 3 seconds when run in
isolation (most of that time spent idling), but this should make the
test friendlier to a loaded slow machine.

Fixes: 782ad5b2
Reported-by: Richard W.M. Jones <rjones@redhat.com>
Signed-off-by: Eric Blake <eblake@redhat.com>