include: Expose nbdkit version information to public
Internally, our plugins and filters can (and do!) use PACKAGE_VERSION
to populate the .version field.  But this macro is defined in
config.h, which is unsuitable for installation in /usr/include, so
external plugin authors cannot use it.  It is worth letting our public
interface include a version designation (ideally, users should NOT be
basing compile-time decisions solely on what version they are
compiling against, but instead using actual feature tests, but we
weren't even allowing for a version check).

The addition is done in such as way as to continue to allow version
bumps to touch a single place in configure.ac and have that propagate
to all affected files.

Signed-off-by: Eric Blake <eblake@redhat.com>