server: Cache per-connection can_extents
We want the semantics of a default 'everything is allocated' response
to .extents for any filter calling into a backend, not just at the top
layer.  To do that, we need to move the fallback logic into backend.c,
and to cache the results of .can_extents.  Without this patch, once
one backend has nothing to report other than allocated, all other
filters were unable to provide allocation details - although it turns
out that the only filter that provides more than a subset of the
underlying plugin is the truncate filter (perhaps the cow filter could
be taught to defer to the underlying plugin for portions that have not
yet been modified, but that's a bigger patch).

$ cat script
case "$1" in
    get_size) echo 1m;;
    *) exit 2 ;;
esac

Pre-patch:

$ ./nbdkit -U - -r --filter=truncate sh script truncate=2m \
  --run 'qemu-img map -f raw --output=json $uri'
[{ "start": 0, "length": 2097152, "depth": 0, "zero": false, "data": true, "offset": 0}]

Post-patch:
$ ./nbdkit -U - -r --filter=truncate sh script truncate=2m \
  --run 'qemu-img map -f raw --output=json $uri'
[{ "start": 0, "length": 1048576, "depth": 0, "zero": false, "data": true, "offset": 0},
{ "start": 1048576, "length": 1048576, "depth": 0, "zero": true, "data": false, "offset": 1048576}]

Signed-off-by: Eric Blake <eblake@redhat.com>