server: Create new backend_* functions for central handling
I'm planning to move sanity checking and per-connection caching into
common code (right now, we check and cache things at the connection
level for the final filter, but not at other layers; it would be nicer
to do this at each step through the filters).  This starts the process
by extending the just-created backend.c to add wrappers for all of the
handshake and request callbacks present in a filter's next_ops.  This
lets us consolidate common debugging output (which requires a tweak to
test-layers), and moves assertion checking (such as what was
duplicated in 0b8ad8c8) into a single location.

Of note: .extents no longer returns 0 for a 0-length request made by a
filter (such a request cannot be made by the client, because
protocol.c:valid_range rejects it); commit 5a1db63d showed a case
where a filter ended up doing that accidentally.

Signed-off-by: Eric Blake <eblake@redhat.com>