server: Rework storage of per-backend handle
Right now, each connection maintains a bare list of void* handles
obtained from the backends, which is reallocated on-the-fly as
additional filters go through .open (and where a filter that skips
.open does not necessarily resize the array).  However, we know the
size of the array in advance (because we know how many filters were
loaded) and can avoid realloc churn; and we also want to move the code
towards being able to cache more than just the handle from the
filter/plugin.  So it makes sense to have each connection maintain an
array of structs that includes the handle and anything else associated
with a given backend/connection combination.

It feels a bit asymmetric to have backend_set_handle() but
connection_get_handle(), but it was less churn this way.

This patch refactors the allocation and tracking of handles, while
upcoming patches add some caching made possible by the new array of
structs.

Signed-off-by: Eric Blake <eblake@redhat.com>