server: Consolidate common backend tasks into new backend.c
Both plugin.c and filter.c add the same three fields on top of struct
backend, and use very similar code in initializing them.  Let's stop
the duplication, by moving those three fields into struct backend, and
creating a new backend.c for manipulating them.  In turn, we can drop
the backend->name() accessor in favor of just directly accessing the
name field, and change set_debug_flags() to be static.  This is a net
reduction in lines of code (remember, the diffstat is also counting
comment additions, including license boilerplate).

As part of the consolidation, use ctype functions instead of
open-coding check for alnum characters (although I'm not aware of
anyone trying to port nbdkit on an EBCDIC machine).

Signed-off-by: Eric Blake <eblake@redhat.com>