server: Adjust validate_request() to check eflags
It's slightly cleaner to write our client validation in terms of what
we advertised to the client, rather than what we cached about what our
plugin supports.  The behavior is the same, since what we advertise is
based on the plugin contents; but this rewrite makes for fewer places
that are directly accessing the single per-connection cache points,
for less churn in an upcoming patch changing to instead use
per-backend cache points.

Note that since validate_request() drops any use of CMD_FLAG_FUA when
it was not advertised, the 'fua' variable in handle_request() is not
adding any useful information.

Signed-off-by: Eric Blake <eblake@redhat.com>