filters: Pass readonly through to .prepare
Some filters need to know whether the underlying plugin supports
various write operations (for example, --filter=nozero zeromode=notrim
requires that the underlying can_zero returns true).  However, if the
filter itself is being used in a readonly mode, .can_zero will fail,
but in this situation we know that .zero will never be called, so it
is not worth treating that failure as fatal.  It is already possible
for a filter to know if it is readonly based on the parameter passed
to .open, but this puts a burden on a filter to implement .open just
to store that bit of information; it's just as easy to alter the
.prepare signature to pass the same value back, for less effort on the
filters.

This is an API change, and thus all filters have to be touched; but it
makes it easier to alter a command line by adding -r without causing
the filter to fall apart over the loss of writes to the underlying
plugin.  But thanks to our recent filter API change (commit fe8ad62e),
we don't have to worry about the change being a problem for mismatched
versions.

Signed-off-by: Eric Blake <eblake@redhat.com>