tests: Avoid -Wshadow warnings
In test-layers, declaring a variable 'log' technically conflicts with
POSIX's rule that <math.h> reserves log(), even if we don't use
math.h.  Picking a different name shuts up the compiler.

In test-exit-with-parent, the use of 'pid' for the grandchild process
collides with the (unused) global 'pid' from test.h; but when you
realize that the use of 'nbdpid' in the child process does not collide
with its use in the monitor process, we don't need an extra
declaration.

In web-server, the only caller of xpread passed global 'fd' to a
shadowed name; just rely on the global instead.

Signed-off-by: Eric Blake <eblake@redhat.com>